+++ src/Thebes.jl
@@ -3,6 +3,14 @@ module Thebes
 using Luxor
 using StaticArrays
 using Rotations
+using Requires
+#using LaTexStrings
+
+function __init__()
+    @require MathTeXEngine = "0a4f8689-d25c-4efe-a92b-7142dfc1aa53" begin
+        @require LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f" include("latex.jl")
+    end
+end
 
 include("Point3D.jl")
 include("utils.jl")
@@ -39,5 +47,6 @@ export project, Projection, newprojection,
        text3D
 
 const CURRENTPROJECTION = [newprojection(Point3D(100, 100, 100), Point3D(0, 0, 0), Point3D(0, 0, 1))]
+
 
 end